Reorganize modules and update imports
